requirements.txt
pymysql
flask
faker
cryptography
Mastodon.py
PyYAML
beautifulsoup4